Cleanup: move node startup code into a class that can be sub-classed to customise behaviour.
