+++ node/src/main/kotlin/net/corda/node/SerialFilter.kt
@@ -1,5 +1,6 @@
 package net.corda.node
 
+import net.corda.node.internal.Node
 import java.lang.reflect.Field
 import java.lang.reflect.Method
 import java.lang.reflect.Proxy
@@ -24,7 +25,7 @@ internal object SerialFilter {
         // JDK 8u121 is the earliest JDK8 JVM that supports this functionality.
         filterInterface = getFilterInterface("java.io")
                 ?: getFilterInterface("sun.misc")
-                ?: failStartUp("Corda forbids Java deserialisation. Please upgrade to at least JDK 8u121.")
+                ?: Node.failStartUp("Corda forbids Java deserialisation. Please upgrade to at least JDK 8u121.")
         serialClassGetter = Class.forName("${filterInterface.name}\$FilterInfo").getMethod("serialClass")
         val statusEnum = Class.forName("${filterInterface.name}\$Status")
         undecided = statusEnum.getField("UNDECIDED").get(null)
