+++ test-utils/src/main/kotlin/net/corda/testing/driver/Driver.kt
@@ -19,7 +19,7 @@ import net.corda.core.node.NodeInfo
 import net.corda.core.node.services.ServiceInfo
 import net.corda.core.node.services.ServiceType
 import net.corda.core.utilities.*
-import net.corda.node.LOGS_DIRECTORY_NAME
+import net.corda.node.internal.NodeStartup
 import net.corda.node.services.config.*
 import net.corda.node.services.network.NetworkMapService
 import net.corda.node.services.transactions.RaftValidatingNotaryService
@@ -698,7 +698,7 @@ class DriverDSL(
                         ),
                         jdwpPort = debugPort,
                         extraJvmArguments = extraJvmArguments,
-                        errorLogPath = nodeConf.baseDirectory / LOGS_DIRECTORY_NAME / "error.log",
+                        errorLogPath = nodeConf.baseDirectory / NodeStartup.LOGS_DIRECTORY_NAME / "error.log",
                         workingDirectory = nodeConf.baseDirectory
                 )
             }
